[ADD] myAccount, changePassword, fitur: (alert, confirm)

+++ src/scripts/views/pages/setting.js
@@ -1,13 +1,15 @@
 import Swal from 'sweetalert2';
-import LocalStorageSource from '../../data/localStorageSource';
 import '../../components/setting-page';
 import UserDbSource from '../../data/userDbSource';
 import UrlParser from '../../routes/url-parser';
+import '../../components/loading/loading-page';
+import LocalStorageSource from '../../data/localStorageSource';
 
 const Setting = {
   async render() {
     return `
-            <setting-page></-page>
+            <setting-page></setting-page>
+            <loading-page></loading-page>
         `;
   },
 
@@ -24,17 +26,25 @@ const Setting = {
       skeletonSetting.classList.remove('d-none');
     };
 
+    const loading = document.querySelector('loading-page');
+
+    const removeLoading = () => {
+      loading.classList.add('d-none');
+    };
+
     addSkeleton();
     const user = await UserDbSource.detailUser(url.id);
     if (user) {
-      containerSetting.innerHTML += `
+      containerSetting.innerHTML = `
+                        <img src="images/profile-pic.png" class="bg-light rounded-circle" style="width:80px; height:80px;">
                         <div class="px-3 pt-3">
                             <h5 class="text-light">${user.name}</h5>
-                            <p class="text-light">${user.number}</p>
+                            <p class="text-light">+62 ${user.number}</p>
                         </div>
                         `;
     } else {
-      containerSetting.innerHTML += `
+      containerSetting.innerHTML = `
+                        <img class="bg-secondary rounded-circle" style="width:80px; height:80px;">
                         <div class="px-3 pt-3">
                             <h5 class="text-light">Name is Invalid</h5>
                             <p class="text-light">Number is Invalid</p>
@@ -42,13 +52,23 @@ const Setting = {
                         `;
     }
     removeSkeleton();
+    removeLoading();
 
     const myaccountButton = document.querySelector('#myaccountButton');
 
     myaccountButton.addEventListener('click', () => {
       const idUser = LocalStorageSource.getUserFromLocalStorage();
       if (idUser) {
-        window.location.assign(`#/setting/${idUser || '-'}/account`);
+        window.location.assign(`#/setting/${idUser}/account`);
+      }
+    });
+
+    const changePassword = document.querySelector('#changePassword');
+
+    changePassword.addEventListener('click', () => {
+      const idUser = LocalStorageSource.getUserFromLocalStorage();
+      if (idUser) {
+        window.location.assign(`#/setting/${idUser}/password`);
       }
     });
 
@@ -61,17 +81,29 @@ const Setting = {
       if (window.navigator.onLine) {
         const idUser = LocalStorageSource.getUserFromLocalStorage();
         if (idUser) {
-          window.location.assign('#/');
-          LocalStorageSource.deleteUserFromLocalStorage();
-          settingNav.classList.add('d-none');
-          loginNav.classList.remove('d-none');
-          registerNav.classList.remove('d-none');
           Swal.fire({
-            position: 'center',
-            icon: 'success',
-            title: 'Logout has succed',
-            showConfirmButton: false,
-            timer: 1500,
+            title: 'Are you sure?',
+            text: `You want to logout from account ${user.name}`,
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, Logout!',
+          }).then((result) => {
+            if (result.isConfirmed) {
+              Swal.fire({
+                position: 'center',
+                icon: 'success',
+                title: 'Logout has succed',
+                showConfirmButton: false,
+                timer: 2500,
+              });
+              window.location.assign('#/');
+              LocalStorageSource.deleteUserFromLocalStorage();
+              settingNav.classList.add('d-none');
+              loginNav.classList.remove('d-none');
+              registerNav.classList.remove('d-none');
+            }
           });
         } else {
           Swal.fire({
